requirements.txt
# Part of SkyOps Agent system – see README for architecture

# Google Sheets Integration
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.115.0

# Data Handling
pandas>=2.2.0

# Environment Variables
python-dotenv>=1.0.0

# Streamlit for UI
streamlit>=1.31.0

# LLM Integration (choose one or both)
openai>=1.12.0
anthropic>=0.18.0
